Add some more UI stats about the current number of nodes

+++ src/Device/kernels/IlluminationAwareKDTree/PromoteGuidingCells.h
@@ -11,10 +11,12 @@
 
 #ifndef __KERNELCC__
 GLOBAL_KERNEL_SIGNATURE(void)
-inline IlluminationAwareKDTree_PromoteGuidingCells(IlluminationAwareKDTreeDevice illumination_aware_kd_tree, unsigned int original_guiding_node_count, int x)
+inline IlluminationAwareKDTree_PromoteGuidingCells(IlluminationAwareKDTreeDevice illumination_aware_kd_tree,
+												   unsigned long long int original_guiding_node_count,
+												   int x)
 #else
 GLOBAL_KERNEL_SIGNATURE(void)
-IlluminationAwareKDTree_PromoteGuidingCells(IlluminationAwareKDTreeDevice illumination_aware_kd_tree, unsigned int original_guiding_node_count)
+IlluminationAwareKDTree_PromoteGuidingCells(IlluminationAwareKDTreeDevice illumination_aware_kd_tree, unsigned long long int original_guiding_node_count)
 #endif
 {
 #ifdef __KERNELCC__
